view building code for main controller extracted to separate view class
